requirements.txt
Flask>1
PyYAML
requests
slack_bolt
slack_sdk

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions